requirements.txt
torch==2.6.0+cu124
pandas==2.2.3
numpy==1.26.4
Pillow==11.1.0
OpenCV==4.11.0
Albumentations==2.0.5
tqdm==4.67.1